chore: :wrench: add and configure lint-staged (#4)
* chore(deps): :link: install lint-staged

This update includes the addition of lint-staged version 16.2.6 to both package.json and package-lock.json for improved code quality checks before commits.

* chore: :wrench: add lint-staged configuration for code formatting

This update introduces a lint-staged configuration to automate code formatting with Prettier and ESLint on staged files.

* chore: :wrench: replace pre-commit script with lint-staged

Updated the pre-commit hook to run lint-staged for improved code formatting and linting.

* docs: :books: update README to include lint-staged in Husky setup

Clarified the usage of Husky and lint-staged for running linters and formatters on staged files before commits.

* docs: :books: update README to include lint-staged in Husky setup

Clarified the usage of Husky and lint-staged for running linters and formatters on staged files before commits.